split xds http filters and filter registry
