+++ build_autogenerated.yaml
@@ -1233,7 +1233,8 @@ libs:
   - src/core/xds/grpc/xds_endpoint.h
   - src/core/xds/grpc/xds_health_status.h
   - src/core/xds/grpc/xds_http_fault_filter.h
-  - src/core/xds/grpc/xds_http_filters.h
+  - src/core/xds/grpc/xds_http_filter.h
+  - src/core/xds/grpc/xds_http_filter_registry.h
   - src/core/xds/grpc/xds_http_rbac_filter.h
   - src/core/xds/grpc/xds_http_stateful_session_filter.h
   - src/core/xds/grpc/xds_lb_policy_registry.h
@@ -2039,7 +2040,7 @@ libs:
   - src/core/xds/grpc/xds_endpoint.cc
   - src/core/xds/grpc/xds_health_status.cc
   - src/core/xds/grpc/xds_http_fault_filter.cc
-  - src/core/xds/grpc/xds_http_filters.cc
+  - src/core/xds/grpc/xds_http_filter_registry.cc
   - src/core/xds/grpc/xds_http_rbac_filter.cc
   - src/core/xds/grpc/xds_http_stateful_session_filter.cc
   - src/core/xds/grpc/xds_lb_policy_registry.cc
