+++ gRPC-C++.podspec
@@ -1342,7 +1342,8 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_endpoint.h',
                       'src/core/xds/grpc/xds_health_status.h',
                       'src/core/xds/grpc/xds_http_fault_filter.h',
-                      'src/core/xds/grpc/xds_http_filters.h',
+                      'src/core/xds/grpc/xds_http_filter.h',
+                      'src/core/xds/grpc/xds_http_filter_registry.h',
                       'src/core/xds/grpc/xds_http_rbac_filter.h',
                       'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                       'src/core/xds/grpc/xds_lb_policy_registry.h',
@@ -2620,7 +2621,8 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_endpoint.h',
                               'src/core/xds/grpc/xds_health_status.h',
                               'src/core/xds/grpc/xds_http_fault_filter.h',
-                              'src/core/xds/grpc/xds_http_filters.h',
+                              'src/core/xds/grpc/xds_http_filter.h',
+                              'src/core/xds/grpc/xds_http_filter_registry.h',
                               'src/core/xds/grpc/xds_http_rbac_filter.h',
                               'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
