+++ gRPC-Core.podspec
@@ -2102,8 +2102,9 @@ Pod::Spec.new do |s|
                       'src/core/xds/grpc/xds_health_status.h',
                       'src/core/xds/grpc/xds_http_fault_filter.cc',
                       'src/core/xds/grpc/xds_http_fault_filter.h',
-                      'src/core/xds/grpc/xds_http_filters.cc',
-                      'src/core/xds/grpc/xds_http_filters.h',
+                      'src/core/xds/grpc/xds_http_filter.h',
+                      'src/core/xds/grpc/xds_http_filter_registry.cc',
+                      'src/core/xds/grpc/xds_http_filter_registry.h',
                       'src/core/xds/grpc/xds_http_rbac_filter.cc',
                       'src/core/xds/grpc/xds_http_rbac_filter.h',
                       'src/core/xds/grpc/xds_http_stateful_session_filter.cc',
@@ -3395,7 +3396,8 @@ Pod::Spec.new do |s|
                               'src/core/xds/grpc/xds_endpoint.h',
                               'src/core/xds/grpc/xds_health_status.h',
                               'src/core/xds/grpc/xds_http_fault_filter.h',
-                              'src/core/xds/grpc/xds_http_filters.h',
+                              'src/core/xds/grpc/xds_http_filter.h',
+                              'src/core/xds/grpc/xds_http_filter_registry.h',
                               'src/core/xds/grpc/xds_http_rbac_filter.h',
                               'src/core/xds/grpc/xds_http_stateful_session_filter.h',
                               'src/core/xds/grpc/xds_lb_policy_registry.h',
