+++ grpc.gemspec
@@ -1989,8 +1989,9 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/xds/grpc/xds_health_status.h )
   s.files += %w( src/core/xds/grpc/xds_http_fault_filter.cc )
   s.files += %w( src/core/xds/grpc/xds_http_fault_filter.h )
-  s.files += %w( src/core/xds/grpc/xds_http_filters.cc )
-  s.files += %w( src/core/xds/grpc/xds_http_filters.h )
+  s.files += %w( src/core/xds/grpc/xds_http_filter.h )
+  s.files += %w( src/core/xds/grpc/xds_http_filter_registry.cc )
+  s.files += %w( src/core/xds/grpc/xds_http_filter_registry.h )
   s.files += %w( src/core/xds/grpc/xds_http_rbac_filter.cc )
   s.files += %w( src/core/xds/grpc/xds_http_rbac_filter.h )
   s.files += %w( src/core/xds/grpc/xds_http_stateful_session_filter.cc )
