+++ package.xml
@@ -1971,8 +1971,9 @@
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_health_status.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_fault_filter.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_fault_filter.h" role="src" />
-    <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_filters.cc" role="src" />
-    <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_filters.h" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_filter.h" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_filter_registry.cc" role="src" />
+    <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_filter_registry.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_rbac_filter.cc" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_rbac_filter.h" role="src" />
     <file baseinstalldir="/" name="src/core/xds/grpc/xds_http_stateful_session_filter.cc" role="src" />
