+++ src/core/xds/grpc/xds_http_filter.h
@@ -14,23 +14,16 @@
 // limitations under the License.
 //
 
-#ifndef GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTERS_H
-#define GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTERS_H
+#ifndef GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTER_H
+#define GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTER_H
 
-#include <map>
-#include <memory>
 #include <string>
-#include <utility>
-#include <vector>
 
-#include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "absl/strings/str_cat.h"
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
 #include "upb/reflection/def.h"
-
-#include <grpc/support/port_platform.h>
 
 #include "src/core/lib/channel/channel_args.h"
 #include "src/core/lib/channel/channel_fwd.h"
@@ -127,60 +120,6 @@ class XdsHttpFilterImpl {
   virtual bool IsTerminalFilter() const { return false; }
 };
 
-class XdsHttpRouterFilter final : public XdsHttpFilterImpl {
- public:
-  absl::string_view ConfigProtoName() const override;
-  absl::string_view OverrideConfigProtoName() const override;
-  void PopulateSymtab(upb_DefPool* symtab) const override;
-  absl::optional<FilterConfig> GenerateFilterConfig(
-      const XdsResourceType::DecodeContext& context, XdsExtension extension,
-      ValidationErrors* errors) const override;
-  absl::optional<FilterConfig> GenerateFilterConfigOverride(
-      const XdsResourceType::DecodeContext& context, XdsExtension extension,
-      ValidationErrors* errors) const override;
-  void AddFilter(InterceptionChainBuilder& /*builder*/) const override {}
-  const grpc_channel_filter* channel_filter() const override { return nullptr; }
-  absl::StatusOr<ServiceConfigJsonEntry> GenerateServiceConfig(
-      const FilterConfig& /*hcm_filter_config*/,
-      const FilterConfig* /*filter_config_override*/) const override {
-    // This will never be called, since channel_filter() returns null.
-    return absl::UnimplementedError("router filter should never be called");
-  }
-  bool IsSupportedOnClients() const override { return true; }
-  bool IsSupportedOnServers() const override { return true; }
-  bool IsTerminalFilter() const override { return true; }
-};
-
-class XdsHttpFilterRegistry final {
- public:
-  explicit XdsHttpFilterRegistry(bool register_builtins = true);
-
-  // Not copyable.
-  XdsHttpFilterRegistry(const XdsHttpFilterRegistry&) = delete;
-  XdsHttpFilterRegistry& operator=(const XdsHttpFilterRegistry&) = delete;
-
-  // Movable.
-  XdsHttpFilterRegistry(XdsHttpFilterRegistry&& other) noexcept
-      : owning_list_(std::move(other.owning_list_)),
-        registry_map_(std::move(other.registry_map_)) {}
-  XdsHttpFilterRegistry& operator=(XdsHttpFilterRegistry&& other) noexcept {
-    owning_list_ = std::move(other.owning_list_);
-    registry_map_ = std::move(other.registry_map_);
-    return *this;
-  }
-
-  void RegisterFilter(std::unique_ptr<XdsHttpFilterImpl> filter);
-
-  const XdsHttpFilterImpl* GetFilterForType(
-      absl::string_view proto_type_name) const;
-
-  void PopulateSymtab(upb_DefPool* symtab) const;
-
- private:
-  std::vector<std::unique_ptr<XdsHttpFilterImpl>> owning_list_;
-  std::map<absl::string_view, XdsHttpFilterImpl*> registry_map_;
-};
-
 }  // namespace grpc_core
 
-#endif  // GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTERS_H
+#endif  // GRPC_SRC_CORE_XDS_GRPC_XDS_HTTP_FILTER_H
