+++ src/core/xds/grpc/xds_http_stateful_session_filter.cc
@@ -42,7 +42,7 @@
 #include "src/core/util/upb_utils.h"
 #include "src/core/xds/grpc/xds_common_types.h"
 #include "src/core/xds/grpc/xds_common_types_parser.h"
-#include "src/core/xds/grpc/xds_http_filters.h"
+#include "src/core/xds/grpc/xds_http_filter.h"
 
 namespace grpc_core {
