+++ test/core/xds/xds_http_filters_test.cc
@@ -14,8 +14,6 @@
 // limitations under the License.
 //
 
-#include "src/core/xds/grpc/xds_http_filters.h"
-
 #include <string>
 #include <utility>
 #include <vector>
@@ -50,6 +48,8 @@
 #include "src/core/lib/iomgr/error.h"
 #include "src/core/util/json/json_writer.h"
 #include "src/core/xds/grpc/xds_bootstrap_grpc.h"
+#include "src/core/xds/grpc/xds_http_filter.h"
+#include "src/core/xds/grpc/xds_http_filter_registry.h"
 #include "src/core/xds/xds_client/xds_client.h"
 #include "src/proto/grpc/testing/xds/v3/address.pb.h"
 #include "src/proto/grpc/testing/xds/v3/cookie.pb.h"
