+++ tools/doxygen/Doxyfile.c++.internal
@@ -2994,8 +2994,9 @@ src/core/xds/grpc/xds_health_status.cc \
 src/core/xds/grpc/xds_health_status.h \
 src/core/xds/grpc/xds_http_fault_filter.cc \
 src/core/xds/grpc/xds_http_fault_filter.h \
-src/core/xds/grpc/xds_http_filters.cc \
-src/core/xds/grpc/xds_http_filters.h \
+src/core/xds/grpc/xds_http_filter.h \
+src/core/xds/grpc/xds_http_filter_registry.cc \
+src/core/xds/grpc/xds_http_filter_registry.h \
 src/core/xds/grpc/xds_http_rbac_filter.cc \
 src/core/xds/grpc/xds_http_rbac_filter.h \
 src/core/xds/grpc/xds_http_stateful_session_filter.cc \
